move store and store types to seperate directory
