+++ electron-app/src/renderer/pages/clusterManagement/bottomBar.tsx
@@ -1,50 +1,37 @@
 import { Button, Stack, Typography } from '@mui/material'
 import { observer } from 'mobx-react-lite'
-import { useCallback, useEffect, useState } from 'react'
-import * as clusterMetadataStore from './clusterMetadataStore'
-import * as ClusterMetadataRequester from '../../components/clusterMetadataRequester'
-import * as ClusterRegisterRequester from '../../components/clusterRegisterRequester'
+import { useCallback, useMemo } from 'react'
+import ClusterManagementUIStore from './UIStore/ClusterManagementUIStore'
+import { useResolve } from '../../hooks/container'
+import ClusterMetadataStore from '../../store/clusterMetadataStore'
+import ClusterRegisterStore from '../../store/clusterRegisterStore'
 import browserLogger from '../../logger/browserLogger'
 
 export default observer(function BottomBar() {
-  const store = clusterMetadataStore.useStore()
-  const metadataRequester = ClusterMetadataRequester.useContext()
-  const registerRequester = ClusterRegisterRequester.useContext()
+  const store = useResolve(ClusterManagementUIStore)
+  const metadataStore = useResolve(ClusterMetadataStore)
+  const registerStore = useResolve(ClusterRegisterStore)
 
-  // QUESTION: I'm using mobX for (reactive) state machine, should I use react context + mobX store?
-  // share requester globally or local?
-  // const requester = ClusterRegisterRequester.useStore()
-
-  const [isProcessing, setIsProcessing] = useState(false)
+  const isProcessing = useMemo(() => {
+    return metadataStore.state !== 'ready' || registerStore.state !== 'ready'
+  }, [metadataStore.state, registerStore.state])
 
   const onRegisterAllClicked = useCallback(() => {
-    ;(async () => {
-      // clear items after request
-      const { selectedItems } = store
-      store.resetSelection()
+    // clear items after request
+    const { selectedItems } = store
+    store.resetSelection()
 
-      await registerRequester.request(
+    // NOTE: should this exists in here?
+    registerStore
+      .request(
         selectedItems.map((item) => ({
           clusterName: item.data.metadata.clustername,
           accountId: item.data.metadata.credresolverid,
         }))
       )
-
-      await metadataRequester.fetchMetadata()
-    })()
-  }, [metadataRequester, registerRequester, store])
-
-  useEffect(() => {
-    if (metadataRequester.state !== 'ready') {
-      return setIsProcessing(true)
-    }
-
-    if (registerRequester.state === 'processing') {
-      return setIsProcessing(true)
-    }
-
-    return setIsProcessing(false)
-  }, [metadataRequester.state, registerRequester.state])
+      .then(() => metadataStore.fetchMetadata())
+      .catch(browserLogger.error)
+  }, [metadataStore, registerStore, store])
 
   return (
     <Stack direction="row" width="100%" alignItems="center" gap="16px" marginLeft="32px">
@@ -53,11 +40,6 @@ export default observer(function BottomBar() {
       </Button>
       {/* on ready (selecting items...) */}
       <Typography>{store.selection.count} Clusters Selected.</Typography>
-      {isProcessing && (
-        <Typography>
-          processing {registerRequester.processedCount} / {registerRequester.length}
-        </Typography>
-      )}
     </Stack>
   )
 })
