+++ electron-app/src/renderer/pages/clusterManagement/clusterInfoList.tsx
@@ -6,7 +6,7 @@ import { useCallback, useMemo, useState } from 'react'
 import LINQ from 'linq'
 import { toJS } from 'mobx'
 import { container } from 'tsyringe'
-import { ClusterMetadataItem, useStore } from './clusterMetadataStore'
+import { ClusterMetadataItem, useStore } from './UIStore/ClusterManagementUIStore'
 import { ClusterInformationStatus } from '../../protos/kubeconfig_service_pb'
 import browserLogger from '../../logger/browserLogger'
 import { useAutorun } from '../../hooks/mobx'
